Fix json dumps error for field lists
Python3 JSON can not dump bytes.
Whenever a submap contains bytes, convert to compatible UTF-8 string.

While we are here, refactor mitigations for FieldList types.

Comment in the middle:
 # This causes a critical error if dntoeu is defined
   # for i in v:
   #     print(i)

Was discovered when attempting to write a string function for the
FieldList class. FieldList is not actually iterable whenever it
contains a dntoeu applied to itself.

If that is ever supported:
   d[k] = str(i)